conversion checked

+++ src/draw_line.c
@@ -6,7 +6,7 @@
 /*   By: mleblanc <mleblanc@student.42quebec.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2021/11/09 13:10:27 by mleblanc          #+#    #+#             */
-/*   Updated: 2021/11/15 21:38:13 by mleblanc         ###   ########.fr       */
+/*   Updated: 2021/12/01 07:07:36 by mleblanc         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,12 +30,12 @@ static void	plotline_low(t_buffer *buf, t_vec2 p0, t_vec2 p1, uint32_t c)
 	int32_t	yi;
 	int32_t	d;
 
-	delta = (t_vec2i){p1.x - p0.x, p1.y - p0.y};
+	delta = (t_vec2i){(int32_t)(p1.x - p0.x), (int32_t)(p1.y - p0.y)};
 	yi = 1;
 	if (delta.y < 0)
 		flip_sign(&yi, &delta.y);
 	d = (delta.y * 2) - delta.x;
-	it = (t_vec2i){p0.x, p0.y};
+	it = (t_vec2i){(int32_t)p0.x, (int32_t)p0.y};
 	while (it.x < (int32_t)p1.x)
 	{
 		put_pixel(buf, it.x, it.y, c);
@@ -57,12 +57,12 @@ static void	plotline_high(t_buffer *buf, t_vec2 p0, t_vec2 p1, uint32_t c)
 	int32_t	xi;
 	int32_t	d;
 
-	delta = (t_vec2i){p1.x - p0.x, p1.y - p0.y};
+	delta = (t_vec2i){(int32_t)(p1.x - p0.x), (int32_t)(p1.y - p0.y)};
 	xi = 1;
 	if (delta.x < 0)
 		flip_sign(&xi, &delta.x);
 	d = (2 * delta.x) - delta.y;
-	it = (t_vec2i){p0.x, p0.y};
+	it = (t_vec2i){(int32_t)p0.x, (int32_t)p0.y};
 	while (it.y < (int32_t)p1.y)
 	{
 		put_pixel(buf, it.x, it.y, c);
